New PinManager
In contrast to the old PinManager, the new PinManager does not attempt
to collapse pin requests. Each pin corresponds to a sticky flag on a
pool. This change is motivated by:

- Not all read requests use the same read pool (depends on the
  PoolManager configuration). Hence we cannot assume that two pin
  requests should necessarily be effectuated on the same pool.

- If a pool is down, then a new pin request shouldn't be "attached" to the
  existing pin on the pool that is down. Instead the file should be
  marked sticky on a new pool.

- PoolManager already contains logic to collapse read/stage
  requests and it already contains logic that will make it pick a read
  pool deterministically if load permits.

- If a pin request for whatever reason gets stuck (bugs happen) then
  it shouldn't block other pin requests on the same file. The old design
  however suffered from this exact problem.


The new PinManager no longer relies on AuthorizationRecord. Each pin has
an owning uid and gid and authorization is performed against those two
fields.

The new PinManager uses JDO for ORM and LiquiBase for schema management.
It uses Spring for transaction demarcation. AspectJ is used for weaving
transaction aspects into the code.

Unpinning is performed asynchronously, which means that upon an unpin
request the pin record is updated, but the sticky flag is not removed
immediately. It is the job of a periodic task to remove those sticky
flags and delete the pin record in the database. This approach reduces
the latency of processing an unpin request and solves failure handling
in case the pool is currently offline: If the pool is offline, the
background task will retry the operation next time it runs.

Expiration is handled by a periodic task which moves expiration pins to
the UNPINNING state.  The regular periodic unpinning task will do the
rest.

Expiration also covers failure recovery. During pinning or while pins
are moved, the expiration time is set such that if the operation fails
for whatever reason, the pin will eventually expire and any stale sticky
flags will be removed.

Pin lifetime extension is implemented as a special case of pin moving.


Basic unit tests for testing positive cases exist. There is plenty of
generic code for stubbing cell communication that could be reused in
other tests. I do however plan to rewrite large parts of CellStub,
AbstractCell, CellMessageDispatcher etc to make cell communication more
testable. I therefore propose not to move the current stub code out of
the pin manager test.

Todo in future patches:

- Make pin move processor asynchronous (or use a separate thread pool)

- Special handling of disk only files

- Better interaction with PoolManager: Pinning a file is not actually the
  same as reading a file. Hence we must not use PoolMgrSelectReadPoolMsg
  (that message modifies the cost). This also gives us an opportunity to
  pass FileAttributes along to PoolManager and thus to reduce latency.

- Add logic to throttle unpinning. This is to avoid overloading pools or
  pinmanager. Also possibly add a pool blacklist to improve how we deal
  with offline pools. Have a look at the chimera cleaner for how this
  could be done.

- Fix issues with singleton aspect.

Target: trunk
Patch: http://rb.dcache.org/r/2852/
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>