logging: Fix CDC setup in cells
The logging context of a thread must be initialized for use in a
particular cell. This is performed by calling CDC.setCellsContext().

CDC.setCellsContext failed to properly clear the old context: In
particular the NDC and the session survived. This presented a problem
when sending a message from one cell to another cell in the same domain.
In that case the receiving thread would be initialized with session and
ndc information from the sending cell, causing this information to be
duplicated.

This patch fixes the problem by letting CDC reset the session and CDC
when initializing a thread for use with a cell. The patch also renamed
setCellContext to reset and lets the reset method return the previous
context as a CDC instance. This captures the common use case in which we
want to restore the old context at some point.

The apply method is renamed to restore as it better captures the meaning
of that method.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2899/