PoolManager: Remove file check
PoolManager could optionally verify the correctness of the cache
location information it gets from the PNFS manager. Doing so adds extra
round trips to the pools, slowing down pool selection. Several doors
turned this feature of and would instead retry pool selection in the
rare event of stale cache location entries.

This patch completely removes the file check from PoolManager.

The DCAP door is changed such that the option for controlling the file
check is removed. The DCAP door already contains logic to retry pool
selection if needed.

The WebDAV door is updated to retry pool selection if needed.

Stager and CopyManager do not have retry logic. However the failure they
get in case of stale entries is however a failure they had to deal with
before this patch as the information may become out of date between pool
manager performs the selection and the door submits the mover. SRM
contains high level retry logic so I do not consider the loss of this
check critical for Stager and CopyManager.


Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2923/