cells: make CellEvent immutable and thread-safe
In the cells package, the CellEvent is the base class for the various events a cell's 
nucleus may encounter; this includes the receipt of an incoming message.  The 
CellEvent class wraps some target object (a CellMessage in the case of 
communication).  The target object is mutable but without any synchronisation.  This 
prevents thread-safe operations with CellEvent objects without further, external 
synchronisation.

The mutability of CellEvent is unnecessary.  This patch removes the unused setSource 
method and marks the wrapped object and type fields as final.  This makes CellEvent 
thread-safe.

Patch: http://rb.dcache.org/r/2949/
Target: trunk
Acked-by: Tigran Mkrtchyan