Security: protect against possible NPE
The GSISelectChannelEndPoint class contains a method 'flush' that is confused about 
whether one of its arguments, Buffer header, may be null.  There is an initial dereference 
without checking if the Buffer is null followed by a couple of dereferences that are 
protected against header being null.

The current code-base does not call flush with a null first argument; however, the JavaDoc 
for the class GSISelectChannelEndPoint extends (Jetty's SelectChannelEndPoint) makes no 
claims about the arguments, one way or the other.

To be safe, this patch fixes the confusion by allowing the possibility of Buffer argument 
header being null.  It protects the initial dereference so that we don't throw NPE.

Target: trunk
Patch: http://rb.dcache.org/r/2957/
Acked-by: Tirgan Mkrtchyan