httpd: remove reference to deleted class
The web admin pages contains support for mapping a UID back to a name.  This is currently 
configured to use a class diskCacheV111.util.MapInfoFields that webadmin instantiates via 
reflection.

A change, committed to SVN with revision 14460, removes MapInfoFields.  This patch is 
described as "Remove dead code"; however, httpd was configured to use MapInfoFields via the 
-fieldMap argument.

This patch removes the reference to MapInfoFields as the feature isn't useful currently (and 
it's what Tigran wants).  It also downgrades the message emitted in no fieldMap option is 
specified from Warning to Info since httpd works fine without the argument.

Target: trunk
Patch: http://rb.dcache.org/r/2960/
Acked-by: Tigran Mkrtchyan