xrootd: Downgrade protocol version to avoid kXR_locate
Turns out that the xrootd protocol at some point changed how clients
locate a data server (pool). In the old days the client would issue
kXR_open to the redirector (door) and be redirected to the data server.
That is what we implemented in dCache and fits well with how dCache
works.

Since xrootd 2.90, clients use a different protocol: They issue a
kXR_locate request that provides 1 or more locations of the file. The
client will then connect to one or more of those. That's something we
don't implement in dCache, and is something which will be quite hard to
add: dCache cannot currently let the client perform the pool selection.

To make matters worse, xrootd even has a concept of directories having a
location: The client will issue a kXR_locate command to find the
locations of a directory and then issue list commands to those
locations. The concept of the redirector knowing the directory content
doesn't exist in recent versions of the protocol.

The change in the protocol is as such quite clever: It allows the
central services to avoid all round trips to the pools. It becomes the
clients responsibilit to recover from stale location information or
pools that are down (by recontacting the the redirector). It is like
part of the door logic becomes embedded into the client. This part may
eventually be realisable in dCache by letting the "xrootd protocol
engine" in the pool be contactable without prior negotiation with a
door.

The change in directory listing on the other hand is difficult to deal
with in dCache: Directory listings are unknown at pools. We can of
course redirect to a pool and let the pool fetch the list from
PnfsManager, but thats a major detour. Alternatively we can introduce
dedicated "list pools" that the redirector sends the client to (similar
to dcap). The list pool would then "contain" all directories. Still,
that seems like an irritating workaround.

How either change fits with authorization is unknown.


For now this patch simply rolls back the version number of the protocol
we advertise as implemented in dCache. The xrootd client contains this
code in the code for issuing locate commands:

   // Old servers will use what's there
   if (fConnModule->GetServerProtocol() < 0x290) {

Thus is we implement something less than 2.9.0, then the client will not
issue locate commands. Hence this patch changes the version to 2.8.9.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6055
Acked-by: Paul Millar <paul.millar@desy.de>