cells: Avoid reply loop
AbstractCell allows message listeners to intercept message to be
forwarded. AbstractCell will forward the message if the message was not
intercepted. In case forwarding the message fails because the receiver
does not exist (ie NoRouteToCellException), AbstractCell will send this
error back to the requestor.

Messages to be forwarded may however be replies themself. Generating
replies for replies comes with the risk of creating messaging loops.
Therefore this patch updates the forwarding logic such that replies are
not generated for replies.


The error could be triggered in a system with space manager enabled.
Steps to reproduce the error where:

1. Disable read pools for a file in PoolManager
2. Issue FTP read on that file

The PoolManager would now have a waiting request for the file.

3. Kill the FTP client. This causes the FTP door to shut down.

The client cell for the request in PoolManager is now gone.

4. Issue an 'rc failed' command in PoolManager.

PoolManager will now send back a failure through space manager. Space
manager will forward the request, however the receiver no longer exists.
AbstractCell in space manager will send a reply back to PoolManager.
PoolManager will accept the message as a new request. The end result is
that 'rc ls' will list a new request and this time the PoolManager is
listed as the requestor.

If the pool gets enabled again, then the request will be processed,
however delivering the successful reply will bounce too. The net effect
is that

1. The request keeps bouncing until PoolManager is restarted

2. The cpu cost of the pool is incremented for every iteration of the
   bounce loop.

The above scenario can only be reproduced on trunk and 1.9.11. This is
so because space manager in earlier versions didn't use AbstractCell.
The bug can however cause similar problems in other cells and hence I
request merging to 1.9.10 and 1.9.9 too.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2947/