pool: Fix termination in case of failed transfers
A migration job must shut down when the queue is empty and no tasks are
running. When scheduling new tasks it may happen that a file entry is
deleted before the corresponding task is started. This is not considered
an error and the job continues scheduling the next task. If however the
job runs out of tasks at this point, then it failed to consider
termination of the job. The effect was that the job would be empty,
however it stayed in the RUNNING state anyway.

The fix is to check for this case at the end of the schedule method.

The patch also fixes a minor issue about cleaning entries from the
_sizes array when they are no longer needed.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2956/