info: log NoRouteToCell at info level
The info cell sends messages based on it's currently belief on dCache's current 
state.  dCache is a somewhat dynamic system, with cells potentially disappearing 
during the lifetime of dCache.  This means that the info-service may send queries 
to a cell only to receive a NoRouteToCellException as a result.  This is not 
abnormal and shouldn't be logged as a warning.

This patch changes the logging of such occurrences from Warning to Info level.

Target: trunk
Patch: http://rb.dcache.org/r/2961/
Acked-by: Tigran Mkrtchyan