cells: make CellMessage serialised payload final
The CellMessage may contain a serialised version of the message's payload.  This 
field is created during object construction (or left as null) and doesn't change 
over the lifetime of the object; therefore may be marked final.

This patch marks the _messageStream field final.  This affords the compiler some 
opportunity to optimise access to this field.

Target: trunk
Acked-by: Tigran Mkrtchyan
Patch: http://rb.dcache.org/r/2953/