space-manager: don't try to create tables if they already exist
On start-up, the space-manager attempts to create database tables even though in the 
majority of cases, these tables already exist.  This leads to a short burst of meaningless 
error messages.

The problem stems from DBManager's createTable method.  This checks if a table exists and 
throws a SQLException if it does. However, it doesn't provide any means to check if the 
table exists.  Since the Manager class (part of space-manager) has no mechanism to check if 
a table exists, it always attempt to create the table.

This patch factors out the DBManager.createTable method's check whether a table already 
exists and makes this available to the Manager class.  By checking if the table exists prior 
to calling createTable, Manager avoids the SQLException from trying to create a table when 
that table already exists.

Patch: http://rb.dcache.org/r/2941/
Target: trunk
Acked-by: Dmitry Litvintsev