configuration: add possibility to mark properties not-for-services
Some properties should not be assigned within the services context.  It would be helpful if we could mark (or annotate) those properties that have no effect in the services context.  Doing this allows dCache to alert the sys-admin if they have misconfigured their dCache.

This patch introduces the possibility for marking a property as "not-for-services".  This currently has no effect, but a subsequent patch will build on this and alert the sys-admin somehow.

Notice that the patch introduces the possibility for a property to have multiple annotations.  This is needed because a property may be marked deprecated and not-for-services at the same time.

Annotations become a comma-separated list of keywords inside parentheses; for example:

  (deprecated)property.to-be-removed = some value
  (forbidden)property.do-not-set = This property has been replaced by property.other-name
  (deprecated,not-for-services)dcache.user=

Target: trunk
Patch: http://rb.dcache.org/r/2863/
Acked-by: Gerd Behrmann