PnfsManager: fix rare index-out-of-bound error when processing messages without PnfsId
PnfsManager has a number of threads for processing requests.  The pathToThreadGroup 
method chooses which thread will process a request without a PNFS-ID.  This method 
should return a number in the range [0 , _threadGroups).

Unfortunately, it contains a bug.  The Math.abs method is not guaranteed to return a 
number >= 0.  If the argument is Integer.MIN_VALUE then Math.abs returns a negative 
number.  The argument is the hashCode of the path; if this hashCode is 
Integer.MIN_VALUE then pathToThreadGroup will return a negative number, PnfsManager 
will attempt to access an array with a negative index, which will fail.

Assuming String.hashCode is a good hash then there is a 1 in 2^32 likelihood of 
triggering this problem.  String.hashCode isn't a good hash, so the likelihood could be 
much more or much less likely.

This patch fixes the problem.

The problem is also present in supported branches: 1.9.9, 1.9.10 and 1.9.11.

Target: trunk
Request: 1.9.9
Request: 1.9.10
Request: 1.9.11