http: don't flooding log file on shutdown
Currently PoolInfoObserverV3 suffers from two related problems within 
the refresh thread.

The first problem is that, when shutting down, the thread is 
interrupted.  This triggers an InterruptedException, which is caught.  
Unfortunately, catching this thread prevents the thread from being 
marked "interrupted", so preventing the loop from exiting.  As a result, 
the thread will continue to loop when it should have shut down.

The second problem is that, if refresh throws a CacheException or a 
RuntimeException then the sleep is ignored.  If the condition that 
triggered the refresh method to throw an exception persists then a 
tight-loop is created.

Current, on shutdown, both problems combine to create a problem.  The 
thread's interrupt is ignored and, when PoolManager shuts down, the 
refresh method will throw a CacheException.  This results in the refresh 
thread forming a tight-loop, logging an error for each iteration.

This patch fixes both these issues: if refresh method throws an 
exception then the loop will still sleep.  Interrupting the thread will 
trigger the thread to die.

Target: trunk
Patch: http://rb.dcache.org/r/2990
Acked-by: Tigran Mkrtchyan