httpd: fix shutdown
The httpd shutdown sequence currently just closes the ServerSocket.  If 
the serving thread is waiting for an incoming connection, then this 
results in the accept method throwing an SocketException, which gets 
logged (at warning level).

Unfortunately, this seems to be the only way to get a thread blocking on 
an incoming connection to exit from the ServerSocket's accept method.  
Therefore, we're forced to ignore any SocketException we might get.

The patch also adjusts the exception handling into a more natural order 
and reducing the scope of the catch from Exception to IOException.

Target: trunk
Patch: http://rb.dcache.org/r/2993/
Acked-by: Tigran Mkrtchyan