cells: add diagnostics for cells reluctant to shutdown
When a sysadmin indicates their desire for dCache to shutdown, various 
attempts are tried with increasing severity.  The system cell attempts 
to shutdown all other cells, then itself.  When shutting down a cell, 
the system cell politely asks the target cell to shutdown.  The target 
cell is given a second to do so before the next stage kicks in.  In the 
next step, any remaining threads within the nucleus' ThreadGroup are 
interrupted.  The system cell waits up to three seconds for all threads 
in the cell's ThreadGroup to terminate.  If this doesn't happen then a 
warning is emitted and the system cell moves on to the next cell.

Once all cells have been killed, the system cell is killed, which should 
terminating the JVM.  If this doesn't happen within 9 seconds then the 
supervising daemon process forcefully kills the JVM.

It is desirable for a cell to shutdown on the polite request rather than 
having all threads interrupted, as the cell may have a preferred 
shutdown sequence.  Interrupting all the threads may no follow that 
sequence.  This is especially true if the domain takes longer than 9 
seconds to shutdown and the supervising daemon forcefully kills the JVM.

To record such problems, the system-cell logs a warning that the cell 
failed to shutdown in 3 seconds.  Unfortunately, this warning is little 
more than "something didn't shutdown quickly enough" message.  It lacks 
sufficient information to diagnose the problem.

This patch increases the verbosity of the system cell's warning message.  
It adds a summary of all the threads in the cell's ThreadGroup.  Each 
thread is listed as a one-line summary followed by that threads 
stack-trace.  The one-line summary has the thread's name, whether the 
thread is alive ('A' or '-') daemon ('D' or '-') interrupted ('I' or 
'-'), the thread's priority and it's current state.

Here is a sample of the output showing three threads:


05 Feb 2011 19:10:38 (System) [] Timeout waiting for httpd
05 Feb 2011 19:10:38 (System) [] Thread: Listener [A-I] (5) RUNNABLE
05 Feb 2011 19:10:38 (System) []     java.net.PlainSocketImpl.socketAccept(Native Method)
05 Feb 2011 19:10:38 (System) []     java.net.AbstractPlainSocketImpl.accept(AbstractPlainSocketImpl.java:358)
05 Feb 2011 19:10:38 (System) []     java.net.ServerSocket.implAccept(ServerSocket.java:470)
05 Feb 2011 19:10:38 (System) []     java.net.ServerSocket.accept(ServerSocket.java:438)
05 Feb 2011 19:10:38 (System) []     dmg.cells.services.HttpServiceCell.run(HttpServiceCell.java:313)
05 Feb 2011 19:10:38 (System) []     dmg.cells.nucleus.CellNucleus$1.run(CellNucleus.java:654)
05 Feb 2011 19:10:38 (System) []     java.lang.Thread.run(Thread.java:636)
05 Feb 2011 19:10:38 (System) [] Thread: restore-collector [A--] (5) TIMED_WAITING
05 Feb 2011 19:10:38 (System) []     java.lang.Object.wait(Native Method)
05 Feb 2011 19:10:38 (System) []     diskCacheV111.poolManager.HttpPoolMgrEngineV3.run(HttpPoolMgrEngineV3.java:107)
05 Feb 2011 19:10:38 (System) []     dmg.cells.nucleus.CellNucleus$1.run(CellNucleus.java:654)
05 Feb 2011 19:10:38 (System) []     java.lang.Thread.run(Thread.java:636)
05 Feb 2011 19:10:38 (System) [] Thread: Thread-40 [A--] (5) WAITING
05 Feb 2011 19:10:38 (System) []     sun.misc.Unsafe.park(Native Method)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.locks.LockSupport.park(LockSupport.java:186)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.locks.AbstractQueuedSynchronizer$ConditionObject.await(AbstractQueuedSynchronizer.java:2043)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.LinkedBlockingQueue.take(LinkedBlockingQueue.java:386)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.ThreadPoolExecutor.getTask(ThreadPoolExecutor.java:1043)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1103)
05 Feb 2011 19:10:38 (System) []     java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:603)
05 Feb 2011 19:10:38 (System) []     dmg.cells.nucleus.CellNucleus$1.run(CellNucleus.java:654)
05 Feb 2011 19:10:38 (System) []     java.lang.Thread.run(Thread.java:636)

The purpose of this information is to allow easy diagnostic of why a 
cell isn't shutting down without having to increase logging threshold or 
otherwise instrumenting the code.

Target: trunk
Patch: http://rb.dcache.org/r/2995/
Acked-by: Tigran Mkrtchyan