srm: add orderly shutdown methods
The SRM doesn't have an orderly shutdown sequence.  Instead, it relies on 
the SystemCell explicitly interrupting all threads in the cell's 
ThreadGroup.  This works, but is noisy.

This patch adds a stop method to Storage. Calling this method shuts down 
the elements of SRM.

Target: trunk
Patch: http://rb.dcache.org/r/2997/
Acked-by: Dmitry Litvintsev
Acked-by: Tigram Mkrtchyan