configuration: tweak error messages
The warning and error messages provided by the configuration system, 
when a sys-admin attempts to adjust the value of an annotated property 
are sub-optimal.  This patch addresses this.

Messages have a consistent prefix "Property <name>: " so that it is 
clear to which property the message refers.  In general, the phrasing 
has changed to indicate what the sys-admin is expected to do, rather 
than give (somewhat vague) information.

The following are samples of possible messages.

If someone copy-and-pasted a property with the annotation when 
attempting to configure that property.  dCache doesn't start and prints 
the following error:

    Property some.property: annotated assignments are not
    allowed; remove "(deprecated)"


The dcache.conf or the layout file attempts to adjust a deprecated 
property.  dCache starts and the following message is logged:

    Property deprecated.property: please review configuration;
    support for this property will be removed in the future


The dcache.conf file or the layout file attempts to adjust an obsolete 
property.  dCache starts and the following message is logged:

    Property obsolete.property: please remove this assignment;
    it has no effect


The dcache.conf or the layout file attempts to adjust a forbidden 
property.  When the property is declared (forbidden), its value is 
empty.  dCache doesn't start and prints the following error:

    Property forbidden.property: may not be adjusted; this
    property no longer affects dCache


dcache.conf or the layout file attempts to adjust a forbidden property.  
When the property is declared (forbidden), its value is "use 
other.property instead".  dCache doesn't start and prints the following 
error:

    Property forbidden.property: may not be adjusted; use
    other.property instead.

Target: trunk
Patch: http://rb.dcache.org/r/3003/
Acked-by: Gerd Behrmann