info: don't make http engine a cell
The http response engine for the info service creates itself a new cell.  This is 
somewhat overkill as all it really needs is the a CellNucleus with which to query the 
info service.

This patch changes the class' constructor to one that receives the containing cell's 
nucleus.  The info cell then uses this nucleus to sendAndWait for messages to the 
info service.

Apart from removing unnecessary bloat of implementing a complete cell, this patch 
also fixes the shutdown sequence, which wasn't handled correctly by the 
InfoHttpEngine "cell"; for example, the LastMessageEvent is not handled correctly by 
the cell.

Target: trunk
Acked-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/2991/