pool: Add a lock for files processed by migration module
When multiple migration tasks operate on the same file, they may cause
various errors as one updates the state of the source file while the
other task processes the same file. The migration module recovers from
these errors, but the failed attempts create a lot of noise and
unnecesary work.

This patch adds a simple locking mechanism to prevent this: A job
attempts to add an entry to a shared map before starting a new task. If
an entry for that pnfsid already exists, the file is skipped in the
current iteration.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2958/