xrootd: Add logging hooks in xrootd door
Netty logging has to be directed towards SLF4J. This was done in the
pool, but not the door. This patch updates the door to use the same
logging hooks as the pool.

Proposed for merge into stable branches as the logging is otherwise 
incomplete (may seem cosmetic, but it makes debugging easier).

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2971/