xrootd: Add kXR_prepare command
kXR_prepare is an xrootd message that asks the server to prepare a file
for access (eg stage it from tape). The message has been around for a
long time in xrootd, but we have not implemented it before.

The newest release of the CMS software however issues kXR_prepare
commands before it opens files. The consequence is that the software
cannot be used with dCache (or dCache with the software).

This patch adds a dummy implementation: It simply always returns OK. 
Real staging is triggered on open like before. We should of course 
eventually implement the real kXR_prepare behaviour, however our current 
implementation does comply with the spec (which points out that the 
prepare call is only a hint).

This patch is submitted for inclusion into stable branches as a 
workaround for the incompatibility caused by the incompleteness of our 
xrootd implementation.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2972/