PoolManager: Fix compatibility with pools older than 1.9.10
In 1.9.10 PoolCostInfo was changes such that the default queue name is
initialized as a final field from a constant declared in IoQueueManager.
This is so because in 1.9.10 the default queue name is hardcoded in the
pool and cannot be changed. The field was kept because PoolCostInfo is
serialized and older pools did allow the default queue name to be
changed. The intention was that if PoolManager received PoolCostInfo
from an old pool, then the field would have a different value and
PoolManager could thus learn the name of the default queue.

Unfortunately it seems the Java compiler seems to optimize the following
code:

    private final String  _defaultQueueName = IoQueueManager.DEFAULT_QUEUE;

    public String getDefaultQueueName(){ return _defaultQueueName ; }

into this code:

    public String getDefaultQueueName(){ return IoQueueManager.DEFAULT_QUEUE; }

Ie it is inlining the value.

Obviously this will not give us the intended behaviour when PoolCostInfo
is deserialized from an old pool.

The problem resulted in NullPointerException in 1.9.10 PoolManager when
used with 1.9.5 pools that specified a different queue name for the
default queue:

04 Feb 2011 11:53:43 (PoolManager) [gridftp-madhatterDomain-1296813221017 GFTP-m
adhatter-<unknown>-3108 PoolMgrSelectWritePool 0000626BD6337F8B4039964C4F01CCE4C
3BB] Exception in requestSucceeded : java.lang.NullPointerException
java.lang.NullPointerException: null
        at diskCacheV111.poolManager.CostModuleV1.messageToForward(CostModuleV1.
java:324) ~[dcache.jar:na]
        at sun.reflect.GeneratedMethodAccessor49.invoke(Unknown Source) ~[na:na]
        at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAcces
sorImpl.java:25) ~[na:1.6.0_22]
        at java.lang.reflect.Method.invoke(Method.java:597) ~[na:1.6.0_22]
        at org.dcache.cells.ShortReceiver.deliver(CellMessageDispatcher.java:52)
 ~[dcache.jar:na]
        at org.dcache.cells.CellMessageDispatcher.call(CellMessageDispatcher.jav
a:255) ~[dcache.jar:na]
        at diskCacheV111.poolManager.CostModuleV1.messageArrived(CostModuleV1.ja
va:369) ~[dcache.jar:na]
        at diskCacheV111.poolManager.PoolManagerV5$WriteRequestHandler.requestSu
cceeded(PoolManagerV5.java:972) [dcache.jar:na]
        at diskCacheV111.poolManager.PoolManagerV5$WriteRequestHandler.run(PoolM
anagerV5.java:947) [dcache.jar:na]
        at java.lang.Thread.run(Thread.java:662) [na:1.6.0_22]
04 Feb 2011 11:55:00 (lmd) [] Exception in Server receive loop (exiting)
java.net.SocketException: Socket closed
        at java.net.PlainDatagramSocketImpl.receive0(Native Method) ~[na:1.6.0_2
2]
        at java.net.PlainDatagramSocketImpl.receive(PlainDatagramSocketImpl.java
:136) ~[na:1.6.0_22]
        at java.net.DatagramSocket.receive(DatagramSocket.java:725) ~[na:1.6.0_2
2]
        at dmg.cells.services.LocationManager$Server.run(LocationManager.java:32
9) ~[cells.jar:na]
        at dmg.cells.nucleus.CellNucleus$1.run(CellNucleus.java:613) [cells.jar:
na]
        at java.lang.Thread.run(Thread.java:662) [na:1.6.0_22]



I added print statement to confirm that getDefaultQueueName did indeed
return IoQueueManager.DEFAULT_QUEUE rather than the value of
_defaultQueueName. I used a heap dump to verify that _defaultQueueName
indeed had the value of the configured default queue on the pool.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2983/