pool: Make FairQueueAllocation interrupt safe
FairQueueAllocation enqueues each thread in a queue and wait for the
calling thread to be at the head of the queue. It then performs the
allocation and dequeues itself.

There is a bug in the code: The dequeue method assumes the thread is at
the head of the queue. That will be the case when everything goes as
planned. However if the thread is interrupted in waitForTurn then the
wrong thread is dequeued. Once the thread that should have been removed
reaches the head of the queue it will not be dequeued, causing all
requests to block.

"P2P Transfer" prio=10 tid=0x000000001adbe000 nid=0xd33 in Object.wait() [0x000000004723c000]
   java.lang.Thread.State: WAITING (on object monitor)
        at java.lang.Object.wait(Native Method)
        at java.lang.Object.wait(Object.java:502)
        at org.dcache.pool.classic.FairQueueAllocation.waitForTurn(FairQueueAllocation.java:51)
        - locked <0x00002aaab93716f0> (a org.dcache.pool.classic.FairQueueAllocation)
        at org.dcache.pool.classic.FairQueueAllocation.allocate(FairQueueAllocation.java:71)
        at org.dcache.pool.repository.v5.WriteHandleImpl.allocate(WriteHandleImpl.java:154)
        at org.dcache.pool.p2p.Companion.transfer(Companion.java:275)
        at org.dcache.pool.p2p.Acceptor$1.run(Acceptor.java:202)


Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Tanja Baranova <tatjana.baranova@desy.de>
Patch: http://rb.dcache.org/r/2985/