ssh: don't trigger IOException when user disconnects
When the user disconnects, the ObjectStreamShell attempt to shutdown the 
connection in such a fashion that the client doesn't loose any information.  
To achieve this, it first closes the outward half of the TCP connection, 
signalling that no further output will be sent.  It then blocks until the 
client disconnects the input half of the connection.

The ObjectStreamShell drains the incoming data and waits for the client to 
close the input half of the TCP connection using jLine's 
ConsoleReader.readLine method.  Unfortunately, jLine will issue a prompt 
and the output connection has been closed.  This results in an IOException.

This patch fixes the problem by using the engine to shutdown.  The engine 
also needs a minor fix to cope with the situation where the server calls 
the close method after the client has disconnected.

The problem is present in all post-1.9.5 dCache releases and has been 
reported against 1.9.10.

Target: trunk
Patch: http://rb.dcache.org/r/3010/
Acked-by: Tigran Mkrtchyan
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9