configuration: take the value of an obsolete property as error message
Properties may be annotated as deprecated, obsolete and forbidden.  
Currently the value of a forbidden property is taken as a custom error 
message.  If the value is the empty string then a suitable default value is 
provided.

This patch provides the same support for obsolete properties.  The value of 
the property is used as a custom error message and, if the value is the 
empty string then a default error message is used.  The default message is 
the error message that is supplied currently.

The patch also adjusts the error message that is printed if the sys-admin 
attempts to copy an annotation when declaring a property value.

Target: trunk
Acked-by: Gerd Behrmann
Patch: http://rb.dcache.org/r/3022/