httpd: fix problem if http has different cellname
The web interface makes use of a standard cells feature: if there's an 
environment called <cellname>Startup (e.g., "httpdSetup" for cell "httpd") 
then this environment is executed after a cell is started.

The httpd service permits the sys-admin to configure which name the cell 
will run as, using the standard cell.name property.  Unfortunately, this is 
not reflected in the Setup environment name.

This patch fixes this by building the startup environment using the 
cell.name property.

Target: trunk
Acked-by: Tigran Mkrtchyan
Patch: http://rb.dcache.org/r/3011/