commons: fix gauge so it doesn't throw an exception if there's a time-shift
We have a class RequestExecutionTimeGauge (org.dcache.commons.stats) 
that keeps track of some statistical measures of a sample data (e.g., 
execution time).

The update method accepts a new long value.  It checks to see that the 
number isn't less than zero (a negative execution time doesn't make 
sense) and throws IllegalArgumentException if it is.

A negative value may be presented if the system's clock suffers a 
backwards time-shift.  If this happens then the 
IllegalArgumentException is thrown.

The current usage of this class assumes that it never throws an 
exception.  This is particularly a problem for the SRM; where, if there 
is a time-shift, the uncaught Exception propagates to the SOAP-server 
layer, which converts it to an SOAP trap that the client then receives.  
Our client will then terminate the SRM transaction.

This patch fixes the problem by removing the exception.  Instead, the 
code logs (at info-level) that invalid data has been presented and 
discards the data.

Target: trunk
Patch: http://rb.dcache.org/r/2654
Acked-by: Dmitry Litvintsev
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5