dcap: Fix space manager integration and add configurable cell names
Fixes a bug in gsidcap and kerberosdcap: When used with space
manager these doors didn't redirect pool messages through space manager.
This would have caused access latency and retention policy of a space
token to be ignored when written to through dcap.

The patch also ensures that the spacemanager, poolManager and
pnfsManager properties are respected.

The patch doesn't apply to 1.9.5, but the bug is present in that release
too and should be fixed.

Target: trunk
Request: 1.9.11
Request: 1.9.10
Request: 1.9.9
Request: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3036/